add authentication to api

+++ apps/api/src/models/addresses/addresses.resolver.ts
@@ -13,6 +13,12 @@ import { CreateAddressInput } from './dto/create-address.input'
 import { UpdateAddressInput } from './dto/update-address.input'
 import { PrismaService } from 'src/common/prisma/prisma.service'
 import { Cinema } from '../cinemas/entities/cinema.entity'
+import {
+  AllowAuthenticated,
+  GetUser,
+} from 'src/common/decorators/auth/auth.decorator'
+import { GetUserType } from '@booking-org/types'
+import { checkRowLevelPermission } from 'src/common/guards'
 
 @Resolver(() => Address)
 export class AddressesResolver {
@@ -21,8 +27,14 @@ export class AddressesResolver {
     private readonly prisma: PrismaService,
   ) {}
 
+  @AllowAuthenticated('admin', 'manager')
   @Mutation(() => Address)
-  createAddress(@Args('createAddressInput') args: CreateAddressInput) {
+  async createAddress(
+    @Args('createAddressInput') args: CreateAddressInput,
+    @GetUser() user: GetUserType,
+  ) {
+    const managers = await this.addressesService.getManagers(args.cinemaId)
+    checkRowLevelPermission(user, managers)
     return this.addressesService.create(args)
   }
 
@@ -36,13 +48,29 @@ export class AddressesResolver {
     return this.addressesService.findOne(args)
   }
 
+  @AllowAuthenticated('admin', 'manager')
   @Mutation(() => Address)
-  updateAddress(@Args('updateAddressInput') args: UpdateAddressInput) {
+  async updateAddress(
+    @Args('updateAddressInput') args: UpdateAddressInput,
+    @GetUser() user: GetUserType,
+  ) {
+    const managers = await this.addressesService.getManagers(args.cinemaId)
+    checkRowLevelPermission(user, managers)
     return this.addressesService.update(args)
   }
 
+  @AllowAuthenticated('admin', 'manager')
   @Mutation(() => Address)
-  removeAddress(@Args() args: FindUniqueAddressArgs) {
+  async removeAddress(
+    @Args() args: FindUniqueAddressArgs,
+    @GetUser() user: GetUserType,
+  ) {
+    const cinema = await this.prisma.cinema.findFirst({
+      where: { Address: { id: args.where.id } },
+      include: { managers: true },
+    })
+    const managers = cinema.managers.map((manager) => manager.uid)
+    checkRowLevelPermission(user, managers)
     return this.addressesService.remove(args)
   }
