+++ apps/api/src/models/seats/seats.resolver.ts
@@ -14,6 +14,12 @@ import { UpdateSeatInput } from './dto/update-seat.input'
 import { PrismaService } from 'src/common/prisma/prisma.service'
 import { Booking } from '../bookings/entities/booking.entity'
 import { Screen } from '../screens/entities/screen.entity'
+import {
+  AllowAuthenticated,
+  GetUser,
+} from 'src/common/decorators/auth/auth.decorator'
+import { checkRowLevelPermission } from 'src/common/guards'
+import { GetUserType } from '@booking-org/types'
 
 @Resolver(() => Seat)
 export class SeatsResolver {
@@ -22,8 +28,20 @@ export class SeatsResolver {
     private readonly prisma: PrismaService,
   ) {}
 
+  @AllowAuthenticated('admin', 'manager')
   @Mutation(() => Seat)
-  createSeat(@Args('createSeatInput') args: CreateSeatInput) {
+  async createSeat(
+    @Args('createSeatInput') args: CreateSeatInput,
+    @GetUser() user: GetUserType,
+  ) {
+    const screen = await this.prisma.screen.findUnique({
+      where: { id: args.screenId },
+      include: { cinema: { include: { managers: true } } },
+    })
+    checkRowLevelPermission(
+      user,
+      screen.cinema.managers.map((m) => m.uid),
+    )
     return this.seatsService.create(args)
   }
 
@@ -38,12 +56,36 @@ export class SeatsResolver {
   }
 
   @Mutation(() => Seat)
-  updateSeat(@Args('updateSeatInput') args: UpdateSeatInput) {
+  async updateSeat(
+    @Args('updateSeatInput') args: UpdateSeatInput,
+    @GetUser() user: GetUserType,
+  ) {
+    const screen = await this.prisma.screen.findUnique({
+      where: { id: args.screenId },
+      include: { cinema: { include: { managers: true } } },
+    })
+    checkRowLevelPermission(
+      user,
+      screen.cinema.managers.map((m) => m.uid),
+    )
     return this.seatsService.update(args)
   }
 
   @Mutation(() => Seat)
-  removeSeat(@Args() args: FindUniqueSeatArgs) {
+  async removeSeat(
+    @Args() args: FindUniqueSeatArgs,
+    @GetUser() user: GetUserType,
+  ) {
+    const seat = await this.prisma.seat.findUnique({
+      where: { id: args.where.id },
+      include: {
+        screen: { include: { cinema: { include: { managers: true } } } },
+      },
+    })
+    checkRowLevelPermission(
+      user,
+      seat.screen.cinema.managers.map((m) => m.uid),
+    )
     return this.seatsService.remove(args)
   }
